Move the attentionvalue stuff to the only place where it is used.
