+++ opencog/attentionbank/CMakeLists.txt
@@ -1,7 +1,6 @@
 
 
 ADD_LIBRARY (attentionbank
-	AtomAttention.cc
 	AttentionBank.cc
 	ImportanceIndex.cc
 	StochasticImportanceDiffusion.cc
